requirements.txt
Django==2.1.3
djangorestframework==3.9.0
pymongo==3.7.2
python-dotenv==0.9.1
pytz==2018.7